% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_years}
\alias{check_years}
\title{Checks which years are available for a given geography}
\usage{
check_years(boundary_type, geog, year, month, type, tag)
}
\arguments{
\item{boundary_type}{Boundary Type}

\item{geog}{Geographical}

\item{year}{Year}

\item{month}{Month}

\item{type}{Type}

\item{tag}{Tag}
}
\description{
Checks which years are available for a given geography
}
