% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlng_eastnorth.R
\name{convert_lnglat}
\alias{convert_lnglat}
\title{Function to convert Eastings / Northings to Lat / Long}
\usage{
convert_lnglat(df, easting, northing)
}
\arguments{
\item{df}{A dataframe with at least 2 columns}

\item{easting}{Column containing Easting}

\item{northing}{Column containing Northing}
}
\value{
Output is a dataframe identical to the input dataframe with two additional columns, 'Long' and 'Lat'.
}
\description{
\code{convert_lnglat} converts columns containing Eastings and Northings in a dataframe to columns containing Longitude and Latitude
}
\details{
This function takes a dataframe with columns containing Eastings and Northings and adds columns containing Longitudes and Latitudes.
}
\examples{

\dontrun{
convert_lnglat(df, "eastings", "northings")
}

}
