% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sf.R
\name{get_admin_sf}
\alias{get_admin_sf}
\title{Get UK Administrative Boundary Shape Files}
\usage{
get_admin_sf(
  geog,
  year = 2019,
  nations = c("E", "S", "W", "N"),
  type = "BGC",
  crs = 4326
)
}
\arguments{
\item{geog}{Type of administrative boundaries}

\item{year}{Year (\code{2018} or \code{2019})}

\item{nations}{For Which UK nations (\code{c("E","W","S","N")})}

\item{type}{Boundary Clipping (\code{"BGC", "BFC", "BFE" or "BUC"})}

\item{crs}{Coordinate Reference System (ESPG). It is recommend to use 4326 (World Geodetic System) or potentially 27700 (UK OS British National Grid)}
}
\value{
Output is...
}
\description{
\code{get_admin_sf} fetches Administrative Boundary shape files from the ONS Open Geography Portal.
}
\details{
This function...
}
\examples{
get_admin_sf("NAT")

\dontrun{
get_admin_sf("UTLA", nations = c("E","W"))
}
}
