% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sf.R
\name{get_census_sf}
\alias{get_census_sf}
\title{Get England and Wales Census Boundary Shape Files}
\usage{
get_census_sf(
  geog,
  year = 2011,
  nations = c("E", "W"),
  type = "BGC",
  crs = 4326
)
}
\arguments{
\item{geog}{Type of census boundaries}

\item{year}{Year (\code{2011} or \code{2021})}

\item{nations}{For Which UK nations (\code{c("E","W")})}

\item{type}{Boundary Clipping}

\item{crs}{Coordinate Reference System (ESPG). It is recommend to use 4326 (World Geodetic System) or potentially 27700 (UK OS British National Grid)}
}
\value{
Output is...
}
\description{
\code{get_census_sf} fetches Census Boundary shape files from the ONS Open Geography Portal.
}
\details{
This function...England and Wales ONLY
}
\examples{
get_census_sf("MSOA")

\dontrun{
get_census_sf("OA", nations = c("E"))
}
}
