% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sf.R
\name{get_sf}
\alias{get_sf}
\title{Get UK Shape Files}
\usage{
get_sf(
  geog,
  year = 2019,
  nations = c("E", "S", "W", "N"),
  type = "BGC",
  crs = 4326
)
}
\arguments{
\item{geog}{Which geographical}

\item{year}{Which year}

\item{nations}{Which UK nations}

\item{type}{Which boundary type}

\item{crs}{Any Coordinate Reference System ESPG, recommend 4326 (World Geodetic System) or potentially 27700 (UK OS British National Grid)}
}
\value{
Output is...
}
\description{
\code{get_sf} fetches shape files from ONS Open Geography Portal.
}
\details{
This function...
}
\examples{
get_sf("NAT")

\dontrun{
get_sf("UTLA", nations = c("E","W"))
}
}
