% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LAD_1112}
\alias{LAD_1112}
\title{LAD lookup 2011-12}
\format{
An object of class \code{data.frame} with 2 rows and 4 columns.
}
\usage{
LAD_1112
}
\description{
LAD lookup 2011-12
}
\keyword{datasets}
