% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LAD_1213}
\alias{LAD_1213}
\title{LAD lookup 2012-13}
\format{
An object of class \code{data.frame} with 4 rows and 4 columns.
}
\usage{
LAD_1213
}
\description{
LAD lookup 2012-13
}
\keyword{datasets}
