% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LAD_1718}
\alias{LAD_1718}
\title{LAD lookup 2017-18}
\format{
An object of class \code{data.frame} with 1 rows and 4 columns.
}
\usage{
LAD_1718
}
\description{
LAD lookup 2017-18
}
\keyword{datasets}
