% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LAD_1819}
\alias{LAD_1819}
\title{LAD lookup 2018-19}
\format{
An object of class \code{data.frame} with 12 rows and 4 columns.
}
\usage{
LAD_1819
}
\description{
LAD lookup 2018-19
}
\keyword{datasets}
