% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LAD_1920}
\alias{LAD_1920}
\title{LAD lookup 2019-20}
\format{
An object of class \code{data.frame} with 4 rows and 4 columns.
}
\usage{
LAD_1920
}
\description{
LAD lookup 2019-20
}
\keyword{datasets}
