% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lea2019lookup}
\alias{lea2019lookup}
\title{LEA Lookup.}
\format{
A data frame with 3 variables:
\code{UTLA19CD},
\code{LEA19CD} and
\code{UTLA19NM}
}
\usage{
lea2019lookup
}
\description{
LEAs have different codes to ONS UTLAs.
}
\keyword{datasets}
