% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_1112}
\alias{lookup_1112}
\title{Lookup 2011-12}
\format{
An object of class \code{data.frame} with 2 rows and 4 columns.
}
\usage{
lookup_1112
}
\description{
Lookup 2011-12
}
\keyword{datasets}
