% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_1213}
\alias{lookup_1213}
\title{Lookup 2012-13}
\format{
An object of class \code{data.frame} with 4 rows and 8 columns.
}
\usage{
lookup_1213
}
\description{
Lookup 2012-13
}
\keyword{datasets}
