% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_1718}
\alias{lookup_1718}
\title{Lookup 2017-18}
\format{
An object of class \code{data.frame} with 1 rows and 4 columns.
}
\usage{
lookup_1718
}
\description{
Lookup 2017-18
}
\keyword{datasets}
