% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_1819}
\alias{lookup_1819}
\title{Lookup 2018-19}
\format{
An object of class \code{data.frame} with 12 rows and 12 columns.
}
\usage{
lookup_1819
}
\description{
Lookup 2018-19
}
\keyword{datasets}
