% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_1920}
\alias{lookup_1920}
\title{Lookup 2019-20}
\format{
An object of class \code{data.frame} with 4 rows and 8 columns.
}
\usage{
lookup_1920
}
\description{
Lookup 2019-20
}
\keyword{datasets}
