% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_2021}
\alias{lookup_2021}
\title{Lookup 2020-21}
\format{
An object of class \code{data.frame} with 7 rows and 12 columns.
}
\usage{
lookup_2021
}
\description{
Lookup 2020-21
}
\keyword{datasets}
