% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_2223}
\alias{lookup_2223}
\title{Lookup 2022-23}
\format{
An object of class \code{data.frame} with 17 rows and 12 columns.
}
\usage{
lookup_2223
}
\description{
Lookup 2022-23
}
\keyword{datasets}
