% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_funcs.R
\name{plus_year}
\alias{plus_year}
\title{Add another year to the lookup}
\usage{
plus_year(df, year)
}
\arguments{
\item{df}{A data.frame}

\item{year}{Year}
}
\description{
Add another year to the lookup
}
