% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sf.R
\name{read_admin}
\alias{read_admin}
\title{Get UK Administrative Boundaries}
\usage{
read_admin(
  geog,
  year = 2019,
  nations = c("E", "S", "W", "N"),
  type = "BGC",
  crs = 4326
)
}
\arguments{
\item{geog}{Type of administrative boundaries}

\item{year}{Year (\code{2018} or \code{2019})}

\item{nations}{For Which UK nations (\code{c("E","W","S","N")})}

\item{type}{Boundary Clipping (\code{"BGC", "BFC", "BFE" or "BUC"})}

\item{crs}{Coordinate Reference System (ESPG). It is recommend to use 4326 (World Geodetic System) or potentially 27700 (UK OS British National Grid)}
}
\value{
Output is...
}
\description{
\code{read_admin} fetches Administrative Boundaries from the ONS Open Geography Portal.
}
\details{
This function...
}
\examples{
read_admin("NAT")

\dontrun{
read_admin("UTLA", nations = c("E","W"))
}
}
