% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sf.R
\name{read_elec}
\alias{read_elec}
\title{Get UK Electoral Boundaries}
\usage{
read_elec(
  geog,
  year = 2018,
  nations = c("E", "W", "S", "N"),
  type = "BGC",
  crs = 4326
)
}
\arguments{
\item{geog}{Type of electoral boundaries}

\item{year}{Year (\code{2018} or \code{2019})}

\item{nations}{For which UK nations (\code{c("E","W", "S", "N")})}

\item{type}{Boundary Clipping}

\item{crs}{Coordinate Reference System (ESPG). It is recommend to use 4326 (World Geodetic System) or potentially 27700 (UK OS British National Grid)}
}
\value{
Output is...
}
\description{
\code{read_elec} fetches Electoral Boundaries from the ONS Open Geography Portal.
}
\details{
This function...
}
\examples{
read_elec("EU")

\dontrun{
read_elec("WM")
}
}
