% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sf.R
\name{read_nuts}
\alias{read_nuts}
\title{Get UK Eurostat NUTS Boundaries}
\usage{
read_nuts(geog, year = 2018, nations = c("E", "W"), type = "BGC", crs = 4326)
}
\arguments{
\item{geog}{NUTS 1, 2 or 3}

\item{year}{Year (\code{2015} or \code{2018})}

\item{nations}{For Which UK nations (\code{c("E","W","S","N")})}

\item{type}{Boundary Clipping}

\item{crs}{Coordinate Reference System (ESPG). It is recommend to use 4326 (World Geodetic System) or potentially 27700 (UK OS British National Grid)}
}
\value{
Output is...
}
\description{
\code{read_nuts} fetches Eurostat NUTS Boundaries from the ONS Open Geography Portal.
}
\details{
This function...
}
\examples{
read_nuts("NUTS1")

\dontrun{
read_nuts("NUTS2", nations = c("E"))
}
}
