% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf.R
\name{read_sf}
\alias{read_sf}
\title{Get UK boundary shapefiles}
\usage{
read_sf(
  geog,
  year = 2021,
  nations = c("E", "S", "W", "N"),
  type = "BGC",
  crs = 4326
)
}
\arguments{
\item{geog}{Type of boundaries

Administrative Boundaries available:

- Countries/Nations (\code{"NAT"})

- Regions (England ONLY) (\code{"GOR"})

- Upper Tier Local Authorities (\code{"UTLA"})

- Lower Tier Local Authorities (\code{"LAD"})

Census Boundaries available:

- Output Area (\code{"OA"})

- Lower Super Output Area (\code{"LSOA"})

- Middle Super Output Area (\code{"MSOA"})

Electoral Boundaries available:

- Westminster Parliamentary Constituencies (\code{"WM"})

- European Union Parliamentary Constituencies (\code{"EU"})

- Welsh Assembly Constituencies (\code{"WAC"})

- Welsh Assembly Regions (\code{"WAR"})

Eurostat NUTS Boundaries available:

   - NUTS Level 1 (\code{NUTS1})

   - NUTS Level 2 (\code{NUTS2})

   - NUTS Level 3 (\code{NUTS3})}

\item{year}{Year (use \code{available_sf()} to see available years for each set of boundaries)}

\item{nations}{For which UK nations (\code{c("E","W","S","N")})}

\item{type}{Boundary clipping

- \code{"BFE"}: Full Resolution Extent of the Realm (Low Tide)

- \code{"BFC"}: Full Resolution Clipped (High Tide)

- \code{"BGC"}: Generalised (20m) Clipped (High Tide)

- \code{"BUC"}: Ultra Generalised (500m) Clipped (High Tide)}

\item{crs}{Coordinate Reference System (ESPG).

It is recommend to use \code{4326} (World Geodetic System) or potentially \code{27700} (UK OS British National Grid)}
}
\value{
Output is a simple feature \code{sf} object containing the relevant boundaries.
}
\description{
\code{read_sf} downloads boundaries from the ONS Open Geography Portal.
}
\examples{
read_sf("NAT")
read_sf("MSOA", year = 2011)
read_sf("EU", year = 2018, type = "BFC")
read_sf("NUTS1", year = 2018, nations = "E")

\dontrun{
read_sf("UTLA", nations = c("E","W"))
read_elec("WM")
}

}
