% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scrape_links}
\alias{scrape_links}
\title{Extract link texts and urls from a web page}
\usage{
scrape_links(url)
}
\arguments{
\item{url}{character an url}
}
\value{
a data frame of link text and urls
}
\description{
Simple function based on [Gist by Paul Rougieux](https://gist.github.com/paulrougieux/e1ee769577b40cd9ed9db7f75e9a2cc2)
}
