% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_sf}
\alias{select_sf}
\title{Interactively select a boundary file to download}
\usage{
select_sf()
}
\value{
A simple feature \code{sf} object containing the interactively selected boundaries.
}
\description{
Interactively select a boundary file to download
}
