% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_url}
\alias{select_url}
\title{Select the url to download from}
\usage{
select_url(boundary_type, geog, year, month, type, crs, tag, num)
}
\arguments{
\item{boundary_type}{Boundary Type}

\item{geog}{Geographical}

\item{year}{Year}

\item{month}{Month}

\item{type}{Type}

\item{crs}{CRS}

\item{tag}{Tag}

\item{num}{Number}
}
\description{
Select the url to download from
}
