% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UTLA_1920}
\alias{UTLA_1920}
\title{UTLA lookup 2019-20}
\format{
An object of class \code{data.frame} with 1 rows and 4 columns.
}
\usage{
UTLA_1920
}
\description{
UTLA lookup 2019-20
}
\keyword{datasets}
