% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BASE_2011}
\alias{BASE_2011}
\title{2011 Lookup between LAD & UTLA}
\format{
A `data.frame` with 4 variables:
\code{LAD11CD},
\code{LAD11NM},
\code{UTLA11CD} and
\code{UTLA11NM}
}
\usage{
BASE_2011
}
\description{
Used as a base dataset
}
\keyword{datasets}
