% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlng_eastnorth.R
\name{convert_lnglat}
\alias{convert_lnglat}
\title{Function to convert Eastings / Northings to Lat / Long}
\usage{
convert_lnglat(df, easting, northing)
}
\arguments{
\item{df}{A dataframe}

\item{easting}{Column containing Easting}

\item{northing}{Column containing Northing}
}
\value{
Output is...
}
\description{
\code{convert_lnglat} converts columns containing Eastings and Northings in a dataframe to columns containing Longitude and Latitude
}
\details{
This function...
}
\examples{

\dontrun{
convert_lnglat(df, "eastings", "northings")
}
}
