% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_year-func.R
\name{plus_year}
\alias{plus_year}
\title{Move to the next year}
\usage{
plus_year(df, geog, year)
}
\arguments{
\item{df}{A data.frame}

\item{geog}{Geographic location identifier, can be mutiple}

\item{year}{Year}
}
\description{
Move to the next year
}
