% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sf.R
\name{read_admin}
\alias{read_admin}
\title{Get UK Administrative Boundaries}
\usage{
read_admin(
  geog,
  year = 2019,
  nations = c("E", "S", "W", "N"),
  type = "BGC",
  crs = 4326
)
}
\arguments{
\item{geog}{Type of administrative boundaries (\code{"NAT", "GOR", "UTLA" or "LAD"})}

\item{year}{Year (\code{2018} or \code{2019})}

\item{nations}{For Which UK nations (\code{c("E","W","S","N")})}

\item{type}{Boundary Clipping

- \code{"BFE"}: Full Resolution Extent of the Realm (Low Tide)

- \code{"BFC"}: Full Resolution Clipped (High Tide)

- \code{"BGC"}: Generalised (20m) Clipped (High Tide)

- \code{"BUC"}: Ultra Generalised (500m) Clipped (High Tide)}

\item{crs}{Coordinate Reference System (ESPG).

It is recommend to use \code{4326} (World Geodetic System) or potentially \code{27700} (UK OS British National Grid)}
}
\value{
Output is a simple feature object containing the relevant boundaries.
}
\description{
\code{read_admin} downloads Administrative Boundaries from the ONS Open Geography Portal.
}
\details{
Boundaries included:

- Countries/Nations (\code{"NAT"})

- Regions (England ONLY) (\code{"GOR"})

- Upper Tier Local Authorities (\code{"UTLA"})

- Lower Tier Local Authorities (\code{"LAD"})
}
\examples{
read_admin("NAT")

\dontrun{
read_admin("UTLA", nations = c("E","W"))
}

}
