% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sf.R
\name{read_census}
\alias{read_census}
\title{Get England and Wales Census Boundaries}
\usage{
read_census(geog, year = 2011, nations = c("E", "W"), type = "BGC", crs = 4326)
}
\arguments{
\item{geog}{Type of census boundaries}

\item{year}{Year (\code{2001} or \code{2011})}

\item{nations}{For Which UK nations (\code{c("E","W")})}

\item{type}{Boundary Clipping}

\item{crs}{Coordinate Reference System (ESPG). It is recommend to use 4326 (World Geodetic System) or potentially 27700 (UK OS British National Grid)}
}
\value{
Output is...
}
\description{
\code{read_census} fetches Census Boundaries from the ONS Open Geography Portal.
}
\details{
This function...England and Wales ONLY
}
\examples{
read_census("MSOA")

\dontrun{
read_census("OA", nations = c("E"))
}
}
