% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sf.R
\name{read_elec}
\alias{read_elec}
\title{Get UK Electoral Boundaries}
\usage{
read_elec(
  geog,
  year = 2018,
  nations = c("E", "W", "S", "N"),
  type = "BGC",
  crs = 4326
)
}
\arguments{
\item{geog}{Type of electoral boundaries (\code{"WM", "EU", "WPC" or "WPR"})}

\item{year}{Year (\code{2018} ONLY)}

\item{nations}{For which UK nations (\code{c("E", "W", "S", "N")})}

\item{type}{Boundary Clipping

- \code{"BFE"}: Full Resolution Extent of the Realm (Low Tide)

- \code{"BFC"}: Full Resolution Clipped (High Tide)

- \code{"BGC"}: Generalised (20m) Clipped (High Tide)

- \code{"BUC"}: Ultra Generalised (500m) Clipped (High Tide)}

\item{crs}{Coordinate Reference System (ESPG).

It is recommend to use \code{4326} (World Geodetic System) or potentially \code{27700} (UK OS British National Grid)}
}
\value{
Output is a simple feature object containing the relevant boundaries.
}
\description{
\code{read_elec} downloads Electoral Boundaries from the ONS Open Geography Portal.
}
\details{
Boundaries included:

- Westminster Parliamentary Constituencies (\code{"WM"})

- European Union Parliamentary Constituencies (\code{"EU"})

- Welsh Assembly Constituencies (\code{"WAC"})

- Welsh Assembly Regions (\code{"WAR"})
}
\examples{
read_elec("EU")

\dontrun{
read_elec("WM")
}

}
