% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sf.R
\name{read_nuts}
\alias{read_nuts}
\title{Get UK Eurostat NUTS Boundaries}
\usage{
read_nuts(
  geog,
  year = 2018,
  nations = c("E", "W", "S", "N"),
  type = "BGC",
  crs = 4326
)
}
\arguments{
\item{geog}{NUTS Level 1, 2 or 3}

\item{year}{Year (\code{2015} or \code{2018})}

\item{nations}{For which UK nations (\code{c("E","W","S","N")})}

\item{type}{Boundary Clipping

- \code{"BFE"}: Full Resolution Extent of the Realm (Low Tide)

- \code{"BFC"}: Full Resolution Clipped (High Tide)

- \code{"BGC"}: Generalised (20m) Clipped (High Tide)

- \code{"BUC"}: Ultra Generalised (500m) Clipped (High Tide)}

\item{crs}{Coordinate Reference System (ESPG).

It is recommend to use \code{4326} (World Geodetic System) or potentially \code{27700} (UK OS British National Grid)}
}
\value{
Output is a simple feature object containing the relevant boundaries.
}
\description{
\code{read_nuts} downloads Eurostat NUTS Boundaries from the ONS Open Geography Portal.
}
\details{
Boundaries include:

- NUTS Level 1 (\code{NUTS1})

- NUTS Level 2 (\code{NUTS2})

- NUTS Level 3 (\code{NUTS3})
}
\examples{
read_nuts("NUTS1")

\dontrun{
read_nuts("NUTS2", nations = c("E"))
}

}
