% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_funcs.R
\name{within_yr_lookup}
\alias{within_yr_lookup}
\title{Create a within year lookup
Simple wrapper for `across_yr_lookup`}
\usage{
within_yr_lookup(year)
}
\arguments{
\item{year}{Year}
}
\description{
Create a within year lookup
Simple wrapper for `across_yr_lookup`
}
